chore: refactor membuilder component/page structure
