+++ src/pages/admin/membership-builder/confirmation.vue
@@ -1,3 +1,10 @@
+<route lang="yaml">
+name: adminMembershipBuilderAccountConfirmation
+meta:
+  pageTitle: membershipBuilder.title
+  role: admin
+</route>
+
 <template>
   <div class="mb-8 grid grid-cols-2 gap-8">
     <div>
@@ -84,24 +91,16 @@
   </div>
 </template>
 <script lang="ts" setup>
-import { onBeforeMount, onBeforeUnmount, ref, watch } from 'vue';
-import { JoinSetupContent } from '../../../../../utils/api/api.interface';
-import { fetchContent, updateContent } from '../../../../../utils/api/content';
-import AppInput from '../../../../forms/AppInput.vue';
-import AppHeading from '../../../../AppHeading.vue';
+import { onBeforeMount, ref } from 'vue';
+import { JoinSetupContent } from '../../../utils/api/api.interface';
+import { fetchContent } from '../../../utils/api/content';
+import AppInput from '../../../components/forms/AppInput.vue';
+import AppHeading from '../../../components/AppHeading.vue';
 import { useI18n } from 'vue-i18n';
-import AppCheckbox from '../../../../forms/AppCheckbox.vue';
-import SetupForm from '../../../join/SetupForm.vue';
-import AuthBox from '../../../../AuthBox.vue';
-import { MembershipBuilderEmitter } from '../membership-builder.interface';
-import useVuelidate from '@vuelidate/core';
-import RichTextEditor from '../../../../rte/RichTextEditor.vue';
-
-const emit = defineEmits(['update:error', 'update:validated']);
-
-const props = defineProps<{
-  emitter: MembershipBuilderEmitter;
-}>();
+import AppCheckbox from '../../../components/forms/AppCheckbox.vue';
+import SetupForm from '../../../components/pages/join/SetupForm.vue';
+import AuthBox from '../../../components/AuthBox.vue';
+import RichTextEditor from '../../../components/rte/RichTextEditor.vue';
 
 const setupContent = ref<JoinSetupContent>();
 
@@ -110,35 +109,7 @@ const { t } = useI18n();
 const stepT = (key: string) =>
   t('membershipBuilder.steps.accountConfirmation.' + key);
 
-const validation = useVuelidate();
-
-watch(validation, () => {
-  emit('update:error', validation.value.$errors.length > 0);
-  emit('update:validated', !validation.value.$invalid);
-});
-
-async function handleUpdate() {
-  if (setupContent.value) {
-    await updateContent('join/setup', setupContent.value);
-  }
-
-  props.emitter.emit('updated');
-}
-
 onBeforeMount(async () => {
-  props.emitter.on('update', handleUpdate);
   setupContent.value = await fetchContent('join/setup');
-
-  watch(
-    setupContent,
-    () => {
-      props.emitter.emit('dirty');
-    },
-    { deep: true }
-  );
-});
-
-onBeforeUnmount(() => {
-  props.emitter.off('update', handleUpdate);
 });
 </script>
